# scca-rest/src/main/java/com/didispace/scca/rest/dto/EncryptKeyDto.java
package com.didispace.scca.rest.dto;

import lombok.Data;
import lombok.NoArgsConstructor;

/**
 * 需要加密的key
 * <p>
 * Created by 程序猿DD/翟永超 on 2018/4/24.
 * <p>
 * Blog: http://blog.didispace.com/
 * Github: https://github.com/dyc87112/
 */
@Data
@NoArgsConstructor
public class EncryptKeyDto {

    private Long id;

    /**
     * 需要加密的key
     **/
    private String eKey;

}